extracted app_state
